feat: economic evaluation implementation
Implemented an evaluation which naively computes an expected yearly gain, then applies both a discount rate and an expected rate of economic growth for a number of years to compute a weighed expected economic gain. This is then compared to the roadbuilding and -maintenance costs, based on which the program makes an evaluation.